correction management.core et bootstrap

+++ org.avm.elementary.management.core/src/org/avm/elementary/management/core/SynchronizeBundlesCommand.java
@@ -242,8 +242,8 @@ class SynchronizeBundlesCommand implements BundleListener {
 					.getAbsoluteFile() + "/" + filename));
 
 			result = (genmd5 != null && md5.equals(genmd5));
-			println("genmd5=" + genmd5 + ", md5=" + md5 + "(check="
-					+ result + ")");
+			println("genmd5=" + genmd5 + ", md5=" + md5 + "(check=" + result
+					+ ")");
 
 		} catch (FileNotFoundException e) {
 			e.printStackTrace();
@@ -298,9 +298,12 @@ class SynchronizeBundlesCommand implements BundleListener {
 				throw new IOException("Cannot get bundle.list from " + url1
 						+ " or " + url2);
 			}
-			
-			File managementDeployedFile = new File( System.getProperty("org.avm.home")	+ "/lib/" + Management.DEPLOYED);
-			boolean managementDeployedFileExist = managementDeployedFile.exists(); 
+
+			File managementDeployedFile = new File(
+					System.getProperty("org.avm.home") + "/lib/"
+							+ Management.DEPLOYED);
+			boolean managementDeployedFileExist = managementDeployedFile
+					.exists();
 
 			updateStartLevel(bundleList);
 			bundleList = getBundlesToUpdate(bundleList);
@@ -393,7 +396,9 @@ class SynchronizeBundlesCommand implements BundleListener {
 			if (changed || managementDeployedFileExist) {
 				try {
 					_management.sendBundleList(_management.getCurrentMode());
-					managementDeployedFile.delete();
+					if (managementDeployedFileExist) {
+						managementDeployedFile.delete();
+					}
 				} catch (Exception e) {
 					println("ERR unable to send bundle list report");
 				}
@@ -410,7 +415,9 @@ class SynchronizeBundlesCommand implements BundleListener {
 						Management.EXITCODE_REBOOT_PLATEFORM);
 			}
 		} catch (UpdateNotAllowedException e1) {
-			println(Constants.SETCOLOR_FAILURE + "/!\\ UPDATE NOT ALLOWED BY SERVER" + Constants.SETCOLOR_NORMAL);
+			println(Constants.SETCOLOR_FAILURE
+					+ "/!\\ UPDATE NOT ALLOWED BY SERVER"
+					+ Constants.SETCOLOR_NORMAL);
 			return;
 		}
 		_timerWaitForEndOfUpdateProcess = new Timer();
@@ -546,7 +553,8 @@ class SynchronizeBundlesCommand implements BundleListener {
 		StringBuffer buffer = new StringBuffer();
 
 		buffer.append(_management.getCurrentDownloadUrl() + "/" + filename);
-		if (_management.getCurrentDownloadUrl().getProtocol().startsWith("http")) {
+		if (_management.getCurrentDownloadUrl().getProtocol()
+				.startsWith("http")) {
 			buffer.append("?mode=" + _management.getCurrentMode());
 		}
 
@@ -758,9 +766,9 @@ class SynchronizeBundlesCommand implements BundleListener {
 		}
 		buf.append(filename);
 
-//		println("downloadURL : " + downloadURL);
-//		println("Filename : " + filename);
-//		println("buf : " + buf.toString());
+		// println("downloadURL : " + downloadURL);
+		// println("Filename : " + filename);
+		// println("buf : " + buf.toString());
 
 		String surl = Utils.formatURL(buf.toString(), false);
 		URL url = new URL(surl);
